// src/types/components/ProjectProps.ts
import { JSX } from 'react'
import ComponentProps from '@/types/components/ComponentProps'
import LinkProps from '@/types/components/LinkProps'

export default interface ProjectProps extends ComponentProps {
  slug?: string
  title: string
  featured?: boolean | false
  description: string | JSX.Element
  category: string
  techStacks: string[]
  otherTechStacks?: string[]
  links?: LinkProps[]
  type:  string
}